Move `PrivacyDropdownMenu` to its own file
